werkt in een keer, whut Veel grote aanpassingen, de basis/skeleton is er nu grotendeels

+++ src/model/Trail.java
@@ -38,7 +38,7 @@ public class Trail extends GameObject {
 		g2d.setComposite(makeTransparent(alpha));
 		
 		g2d.setColor(color);
-		g2d.fillRect(x, y, width, height);
+		g2d.fillRect((int) x, (int) y, width, height);
 		
 		g2d.setComposite(makeTransparent(1));
 	}
